mp3 instead of aac + parallel processing
I realized after a lot of googling that using aac is more trouble
than it is worth, mainly because the codec is proprietary and
because it can hold less metadata compared to mp3 files. I rewrote
the code so that it uses the ffmpeg-python packaget to convert all
files to mp3 with quality > 192kps.

There is no more fussing with the metadata, as ffmpeg automatically
maps all metadata of the original file to the new one.

Furtheremore the script now utilizes joblib to process multiple files
in parallel. The number of jobs can be set with the global variable
WORKERS.